fixed errors

+++ include/utils.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 
 int *load_numbers(const char *path, size_t *count);
-int add_func(int a, int b);
+unsigned long add_func(int a, int b);
 unsigned long max_func(int a, int b);
 
 #endif // UTILS_H
