+++ src/sequential_compute.c
@@ -1,11 +1,11 @@
 #include "utils.h"
 #include <stdio.h>
 
-int sequential_compute(const char *path, int (*func)(int, int)) {
+unsigned long sequential_compute(const char *path, unsigned long (*func)(int, int)) {
     size_t count = 0;
     int *numbers = load_numbers(path, &count);
     if (!numbers || count == 0) return 0;
-    int result = numbers[0];
+    unsigned long result = numbers[0];
     for (size_t i = 1; i < count; ++i) {
         result = func(result, numbers[i]);
     }
